+++ lib/model/video_model.dart
@@ -1,0 +1,4 @@
+class Modeldata2{
+  String?Name2,Image2,video;
+  Modeldata2({this.Name2,this.Image2,this.video});
+}
