+++ pubspec.yaml
@@ -41,6 +41,10 @@ dependencies:
   lottie: ^2.2.0
   provider: ^6.0.5
   image_picker: ^0.8.6+1
+  bottom_bar_matu: ^1.3.0
+  sizer: ^2.0.15
+  video_player: ^2.5.1
+  chewie: ^1.4.0
 
 dev_dependencies:
   flutter_test:
